refactor: replace `io.py` with `input` and `output` packages
I have split the `io.py` file into two package with the following modules:

- `input.py`: functions to handle user inputs
- `panel.py`: `printpnl` and predefined panels
- `print.py`: `print`, `printmd` and output formatting functions
- `color_print.py`: `print_error`, `print_warning`, `print_success` and
`print_info`

The newly introduced `color_print.py` module is to handle the printing for
`file.py` temporarily, involving the `colorama` package.
