+++ requirements.txt
@@ -6,6 +6,7 @@ black==23.1.0
 certifi==2022.12.7
 charset-normalizer==3.1.0
 click==8.1.3
+colorama==0.4.6
 frozenlist==1.3.3
 idna==3.4
 markdown-it-py==2.2.0
